feat: clean up file, document disabled props and split out util
