+++ packages/date-picker/src/utils.ts
@@ -1,0 +1,4 @@
+import { DayOfWeek } from "./DatePicker"
+
+export const daysToNumbers = (days: DayOfWeek[]): number[] =>
+  Array.from(days).map(day => day)
